+++ tomotok/core/inversions/bob.py
@@ -2,7 +2,7 @@
 #
 # Licensed under the EUPL-1.2 or later.
 """
-Contains inversion class for Biorthogonal Basis Decomposition Algorithm using single node basis [1]_ derived from wavelet-vaguelette decomposition [2]_
+Contains inversion class for Biorthogonal Basis Decomposition Algorithm [1]_ derrived from wavelet-vaguelette decomposition [2]_
 
 .. [1] Jordan Cavalier et al 2019 Nucl. Fusion 59 056025
 .. [2] R. Nguyen van yen et al 2011 Nucl. Fusion 52 013005
@@ -17,43 +17,33 @@ class Bob(object):
     """
     BiOrthogonal Basis decomposition
 
-    Parameters
-    ----------
-    dec_mat : scipy.sparse.csr_matrix, optional
-        previsously decomposed matrix, avoids recomputration of decomposition when provided
-    basis : array-like, optional
-        tbd
-    precision : float, optional
-        precision used for normalisation of decomposition matrix, by default 1e-6 
-
     Attributes
     ----------
-    basis : array-like
+    basis : array_like
         :math:`\mathbf{b}_i` basis vectors of reconstruction plane
     dec_mat : scipy.sparse.csr_matrix
         :math:`\hat{\mathbf{e}}_i` decomposed matrix used to transform image into reconstruction plane
     dec_mat_normed : scipy.sparse.csr_matrix
         normalised decomposed matrix
-    precision : float
-        precision used for normalisation of decomposition matrix
-    nnodes : int
-        number of nodes in reconstruction plane
     """
 
-    def __init__(self, dec_mat=None, basis=None, precision=1e-6):
+    def __init__(self, dec_mat=None, basis=None):
+        """
+        Parameters
+        ----------
+        dec_mat : scipy.sparse.csr_matrix, optional
+            previously decomposed matrix, avoids recomputation of decomposition when provided
+        basis : array_like, optional
+            A set of basis vectors used for decomposition
+        """
+        # TODO one parameter holding both dec_mat and basis?
         super().__init__()
         self.basis = basis
-        self.precision = precision
         self.dec_mat = dec_mat
         self.dec_mat_normed = None
-        self.nnodes = None
         return
 
-    def create_basis(self, *args, **kwargs):
-        raise NotImplementedError()
-
-    # TODO add basis as a parameter
-    def decompose(self, gmat):
+    def decompose(self, gmat, basis, reg_factor=0):
         """
         Decomposes the geometry matrix using basis vectors
 
@@ -64,39 +54,40 @@ class Bob(object):
         """
         if gmat.shape[0] < gmat.shape[1]:
             warnings.warn('Biorthogonal algorithm requires more lines of sights than nodes in reconstruction plane to run reliably')
-        self.nnodes = gmat.shape[1]
-        if self.basis is None:
-            self.create_basis()
-        chi = gmat.dot(self.basis)
-        a = chi.T.dot(chi).toarray()
-        b = np.eye(self.nnodes)
+        self.basis = basis
+        image_base = gmat.dot(self.basis)  # e_i previously known as chi, gmat in basis
+        a = image_base.T.dot(image_base).toarray()  # symmetrized geometry matrix in basis
+        if reg_factor:
+            a = a + a.max() * reg_factor * np.eye(*a.shape)
+        b = np.eye(gmat.shape[1])
         res = np.linalg.lstsq(a, b)
-        c = sparse.csr_matrix(res[0])
-        xi = chi.dot(c)
-        self.dec_mat = xi
+        c = sparse.csr_matrix(res[0])  # coefficient matrix
+        self.dec_mat = image_base.dot(c)  # \hat{e}_i previously known as xi, decomposed matrix
         return
 
-    def __call__(self, data, gmat=None, **kw):
+    def __call__(self, data, gmat=None, thresholding=None, **kw):
         """
         Decomposes geometry matrix and projects images
 
         Parameters
         ----------
-        data : np.ndarray
-            contains signals with shape ('#chnl', '#timeslices')
-        gmat : scipy.csr_matrix
+        data : numpy.ndarray
+            contains signals with shape (# channels, # time slices)
+        gmat : scipy.sparse.csr_matrix
             geometry matrix
-        grid : tomotok.io.Pixgrid
-            reconstruction grid
+        thresholding : float, optional
+            if provided, result is processed using thresholding, not performed by default 
         """
         if self.dec_mat is None:
             if gmat is None:
                 raise ValueError('Gmat must be provided for decomposition')
             else:
                 self.decompose(gmat)
-        res = self.dec_mat.T.dot(data)
+        coeffs = self.dec_mat.T.dot(data)  # coordinates in reconstruction basis
+        res = self.basis.dot(coeffs)  # result in node basis
         return res
 
+    # TODO remove? remove normalised parameter?
     def save_decomposition(self, floc, normalised=False):
         """
         Saves 
@@ -109,36 +100,48 @@ class Bob(object):
             selects whether to save normalised matrix, by default False
         """
         floc = str(floc)
-        sparse.save_npz(floc, self.dec_mat, compressed=True)
+        if normalised:
+            sparse.save_npz(floc, self.dec_mat_normed, compressed=True)
+        else:
+            sparse.save_npz(floc, self.dec_mat, compressed=True)
 
-    def normalise(self, precision=None):
-        if precision is None:
-            precision = self.precision
-        xi = self.dec_mat
-        kappa = sparse.linalg.norm(xi, axis=0)
-        idx = kappa > self.precision
+    def normalise(self, precision=1e-6):
+        """
+        Computes normalised decomposition matrix.
+
+        Parameters
+        ----------
+        precision : float, optional
+            neglects decomposition matrix rows with lower norm, by default 1e-6
+        """
+        image_base_adj = self.dec_mat
+        kappa = sparse.linalg.norm(image_base_adj, axis=0)
+        idx = kappa > precision
         norms = np.zeros(kappa.size)
         norms[idx] = (1 / kappa[idx])
-        xi_norm = xi.multiply(sparse.csr_matrix(norms))
+        xi_norm = image_base_adj.multiply(sparse.csr_matrix(norms))
         self.dec_mat_normed = xi_norm
 
-    def thresholding(self, image, c=4, precision=None):
+    # TODO replace image by reconstruction?
+    # TODO return mask only?
+    def thresholding(self, image, c: int, precision: float=1e-6, conv: float=1e-9):
         """
         Applies thresholding method to provided image.
 
         Parameters
         ----------
-        image : np.ndarray
+        image : numpy.ndarray
             flattened image with shape (#pixels, 1)
-        c : int, optional
-            thresholding sensitivity constant, by default 4
+        c : int
+            thresholding sensitivity constant
         precision : float, optional
-            normalisation precision
+            normalisation precision, by default 1e-6
+        conv : float, optional
+            thresholding convergence limit
 
         Returns
         -------
-        np.ndarray
-            thresholded image
+        numpy.ndarray
 
         Raises
         ------
@@ -146,24 +149,95 @@ class Bob(object):
             If tresholding is called before decomposition of geometry matrix
         """
         if self.dec_mat is None:
-            raise RuntimeError('Decompose first, threshold later')
+            raise RuntimeError('Decomposition must be computed prior to thresholding.')
         if self.dec_mat_normed is None:
             self.normalise(precision)
-        temp = sparse.csr_matrix(image.T)
-        a = np.abs(temp.dot(self.dec_mat_normed).toarray())
+        temp = sparse.csr_matrix(image)
+        a = np.abs(self.dec_mat_normed.T.dot(temp).toarray())
 
-        threshold_2 = np.sqrt(c**2 / np.size(a) * a.dot(a.T))
+        threshold_2 = np.sqrt(c**2 / a.size * a.T.dot(a))
         threshold_1 = 0
-        while np.abs(threshold_1-threshold_2) >= 1e-9:
+        while np.abs(threshold_1-threshold_2) >= conv:
             threshold_1 = threshold_2
             a_temp = a[a <= threshold_1]
-            threshold_2 = np.sqrt(c**2 / np.size(a_temp) * a_temp.dot(a_temp.T))
+            threshold_2 = np.sqrt(c**2 / a_temp.size * a_temp.T.dot(a_temp))
 
         out = self.dec_mat.T.dot(image)
-        out[a.T < threshold_2] = 0
+        out = self.basis.dot(out)
+        out[a < threshold_2] = 0
         return out
 
 
 class SimpleBob(Bob):
-    def create_basis(self):
-        self.basis = sparse.eye(self.nnodes)
+    """
+    Automatically creates simple one node basis as an sparse identity matrix
+
+    .. Deprecated:: 1.1
+    """
+
+    def __init__(self, dec_mat=None, basis=None):
+        warnings.warn('SimpleBob is depracated since v1.1', DeprecationWarning)
+        super().__init__(dec_mat, basis)
+    
+    def decompose(self, gmat, basis=None):
+        if basis is not None:
+            warnings.warn('Ignoring basis input')
+        basis = sparse.eye(gmat.shape[1])
+        return super().decompose(gmat, basis)
+
+
+class SparseBob(Bob):
+    """
+    Biorthogonal Basis Decomposition optimized for sparse matrices using inverse matrix calculation.
+    """
+
+    def decompose(self, gmat, basis, reg_factor=0):
+        if gmat.shape[0] < gmat.shape[1]:
+            warnings.warn('Biorthogonal algorithm requires more '
+            'lines of sights than nodes in reconstruction plane to run reliably')
+        self.basis = basis
+        chi = gmat.dot(self.basis)
+        a = chi.T.dot(chi)
+        if reg_factor:
+            a = a + a.max() * reg_factor * sparse.eye(*a.shape)
+        try:
+            c = sparse.linalg.inv(a)
+        except RuntimeError:
+            raise ValueError('Singular symmetrized matrix factor. Try increasing regularisation factor.')
+        self.dec_mat = chi.dot(c)  # xi
+        return
+
+
+class CholmodBob(Bob):
+    """
+    Decomposition optimized for sparse matrices using Cholesky decomposition
+
+    Uses sksparse.cholmod.cholesky to solve the decomposition
+    Requires positive definite symmetrized geometry matrix in reconstruction plane basis.
+    """
+
+    def decompose(self, gmat, basis, reg_factor=1e-3):
+        """
+        Parameters
+        ----------
+        gmat : scipy.sparse.csr_matrix
+        basis
+        reg_factor : float, optional
+            regularisation factor passed to cholesky decomposition
+            determines weight of regularisation by identity matrix relatively to arbitrary matrix maximum value
+        """
+        from sksparse.cholmod import cholesky, CholmodNotPositiveDefiniteError
+        if gmat.shape[0] < gmat.shape[1]:
+            warnings.warn('Biorthogonal algorithm can be prone to failure if there are more '
+            'lines of sights than nodes in reconstruction plane')
+        self.basis = basis
+        chi = gmat.dot(self.basis)
+        a = chi.T.dot(chi)
+        try:
+            factor = cholesky(a, a.max()*reg_factor)
+        except CholmodNotPositiveDefiniteError:
+            raise ValueError('Symmetrized matrix was not positive definite. Try increasing regularisation factor.')
+        b = sparse.csc_matrix(np.eye(gmat.shape[1]))
+        c = factor(b)
+        self.dec_mat = chi.dot(c)  # xi
+        return